backend/requirements.txt
fastapi
uvicorn
gunicorn
google-cloud-resource-manager
google-cloud-datastore
google-cloud-compute
google-cloud-org-policy
google-api-python-client
google-cloud-securitycenter
google-cloud-securitycentermanagement
google-cloud-network-management
google-cloud-dns
protobuf>=4.21.0
python-dotenv
